Organize Zassenhaus error bounds and add term grouper customization
