+++ python/src/qdk_chemistry/algorithms/hamiltonian_unitary_builder/base.py
@@ -7,8 +7,18 @@
 
 from abc import abstractmethod
 
+import numpy as np
+
 from qdk_chemistry.algorithms.base import Algorithm, AlgorithmFactory
-from qdk_chemistry.data import QubitHamiltonian, Settings, UnitaryRepresentation
+from qdk_chemistry.data import (
+    FlatPartition,
+    LayeredPartition,
+    QubitHamiltonian,
+    Settings,
+    TermPartition,
+    UnitaryRepresentation,
+)
+from qdk_chemistry.utils import Logger
 
 __all__: list[str] = [
     "HamiltonianUnitaryBuilder",
@@ -132,6 +142,71 @@ class TimeEvolutionBuilder(HamiltonianUnitaryBuilder):
             UnitaryRepresentation: A UnitaryRepresentation representing the evolution of the given QubitHamiltonian.
 
         """
+
+    def _group_terms(
+        self,
+        qubit_hamiltonian: QubitHamiltonian,
+    ) -> list[list[QubitHamiltonian]]:
+        """Group Hamiltonian terms for decomposition."""
+        partition = qubit_hamiltonian.term_partition
+        if partition is not None:
+            Logger.debug(
+                f"{self.name().capitalize()}: consuming QubitHamiltonian.term_partition "
+                f"(strategy={partition.strategy!r}, num_groups={partition.num_groups})."
+            )
+            return self._groups_from_partition(qubit_hamiltonian, partition)
+
+        Logger.debug(
+            f"{self.name().capitalize()}: no term_partition present; treating each Pauli term as its own group."
+        )
+        return [
+            [
+                QubitHamiltonian(
+                    pauli_strings=[label],
+                    coefficients=[coeff],
+                    encoding=qubit_hamiltonian.encoding,
+                    fermion_mode_order=qubit_hamiltonian.fermion_mode_order,
+                )
+            ]
+            for label, coeff in zip(qubit_hamiltonian.pauli_strings, qubit_hamiltonian.coefficients, strict=True)
+        ]
+
+    def _groups_from_partition(
+        self,
+        qubit_hamiltonian: QubitHamiltonian,
+        partition: TermPartition,
+    ) -> list[list[QubitHamiltonian]]:
+        """Materialise a :class:`TermPartition` into sub-groups."""
+        labels = qubit_hamiltonian.pauli_strings
+        coeffs = qubit_hamiltonian.coefficients
+        encoding = qubit_hamiltonian.encoding
+        fmo = qubit_hamiltonian.fermion_mode_order
+
+        def _make(indices: tuple[int, ...]) -> QubitHamiltonian:
+            return QubitHamiltonian(
+                pauli_strings=[labels[i] for i in indices],
+                coefficients=np.asarray([coeffs[i] for i in indices]),
+                encoding=encoding,
+                fermion_mode_order=fmo,
+            )
+
+        if isinstance(partition, LayeredPartition):
+            layered_groups = partition.groups
+        elif isinstance(partition, FlatPartition):
+            layered_groups = tuple((g,) for g in partition.groups)
+        else:
+            raise TypeError(
+                f"Unsupported TermPartition subtype: {type(partition).__name__}. "
+                "Expected FlatPartition or LayeredPartition."
+            )
+
+        groups: list[list[QubitHamiltonian]] = [
+            [_make(layer) for layer in group_layers if layer] for group_layers in layered_groups
+        ]
+
+        groups = [g for g in groups if g]
+        groups.sort(key=len)
+        return groups
 
 
 class HamiltonianUnitaryBuilderFactory(AlgorithmFactory):
